Add find_bpod_state to BehDat
This ended up moving a method already called find_bpod_state to be renamed find_bpod_state_video. We will see where that function call needs to be patched in when the video pipeline is finalized. The new method is used in response to an optional parameter 'withinState' for the find_bpod_event function, allowing users to retrieve bpod events that occurred within certain states.