Add variable selection screen to startup for AbbasCentral
The old version required the user to have a BehDat array called expSessions, and a metadata file called expMetadata, both created from the function select_experiment. Now to use the app, the user must have an ExpManager variable in their workspace, which stores the sessions and the metadata. This file can be created from the function collect_sessions.