Optimize ppc method
NS6 files are now loaded in segments centered around events rather than loading the entire file into memory. Furthermore, this methodology automatically deals with the 102 samples of zeros that are present in the start of every NS6 file for some reason, which was not accounted for in the previous ppc methodolgy.