Add priorToState param to find_bpod_event
Cleaned up the priorToEvent logic to use the pre-sorted times that I didn't know about before, and use map_bpod_events to give names to the events. An issue that could arise is that sometimes events are what trigger the state transitions, so the event times and start of state time will be identical. I believe that it will not be since when I concatenate events and states together prior to sorting, I add states to the end of the events vector. That way when it is sorted, when times are identical it provides the index that occurred first in the unsorted vector as the first sorted time, which should be the event, and the state as the second time.