Label height of trial sets in raster
The bottom-most trial number is displayed next to the ylines on raster functions that plot multiple trial sets. This could be improved to move the numbering off to the right for readability but as it currently stands is not worth implementing imo. It would take major revisions to the system that copies objects from the figure to the panel.