Alter core param methodology
All files in this commit have had their inputParser scheme adjusted to rely entirely on PresetManager, giving the flexibility to still use name-value pairs if desired but also permitting the use of a single preset. Events are now contained inside the PresetManager object, but can be added to a function call using the name 'event'.