Add BpodParser methods
They are both functional but I'm not sure I want to commit to the current logic. It might be better to break out the subroutines in event_times to different methods, and handle withinState excludeState etc. outside the event_times function. It's kinda sloppy because I wanted to accomodate multiple events with regexp, as well as multiple states for withinState and other options.