Finish state_times for BpodParser
This commit also includes mild changes to find_event and find_bpod_event BehDat methods. PresetManager will be used to include the find_bpod_event conditionals in the upcoming method event_times for BpodParser, these times will be converted to the sampling rate of a BehDat object if it exists.