Fix event_within_state in BpodParser
The output is now identical to the output from find_bpod_event using the 'withinState' flag. The BpodParser uses a PresetManager to populate the intersect fields, so all params that could be used in find_bpod_event should be fully functional once the function event_times is fully tested. For now, only event_within_state is reliable.