Publish find_event_all
This method will soon be renamed find_event, and will be the default function for the previous behavior of both find_event and find_bpod_event. The user should input ('bpod', true) as a name/value pair in the call to find_event_all. It is likely in future updates that calls to this function with Port#[In/Out] will be defaulted to bpod_events.